Фикс: 16-битная маска слишком мала
Для 17 слотов нужна 32-битная маска. Поэтому не прятались пистолеты в новом слоте пистолета, когда это было нужно.

+++ SourcesAXR/xrGame/Actor.h
@@ -679,7 +679,7 @@ protected:
 		void							Check_for_AutoPickUp			();
 		void							SelectBestWeapon				(CObject* O);
 public:
-		void							SetWeaponHideState				(u16 State, bool bSet);
+		void							SetWeaponHideState				(u32 State, bool bSet);
 private://IPhysicsShellHolder
 
 virtual	 void	_BCL	HideAllWeapons					( bool v ){ SetWeaponHideState(INV_STATE_BLOCK_ALL,v); }
