+++ SourcesAXR/xrGameCS/Inventory.h
@@ -108,7 +108,7 @@ public:
 	bool 					IsBeltUseful		() const			{return m_bBeltUseful;}
 	void 					SetBeltUseful		(bool belt_useful)	{m_bBeltUseful = belt_useful;}
 
-	void					SetSlotsBlocked		(u16 mask, bool bBlock);
+	void					SetSlotsBlocked		(u32 mask, bool bBlock);
 
 	void					ChooseItmAnimOrNot	(PIItem pIItem);
 	bool					ItmHasAnim			(PIItem pIItem);
