+++ SourcesAXR/xrGameSoC/Inventory.h
@@ -113,7 +113,7 @@ public:
 	bool 					IsBeltUseful		() const			{return m_bBeltUseful;}
 	void 					SetBeltUseful		(bool belt_useful)	{m_bBeltUseful = belt_useful;}
 
-	void					SetSlotsBlocked		(u16 mask, bool bBlock);
+	void					SetSlotsBlocked		(u32 mask, bool bBlock);
 	TIItemContainer			m_all;
 	TIItemContainer			m_ruck, m_belt;
 	TISlotArr				m_slots;
